feat(mcp): Refactor MCP tool integration and enhance LitMCPSpec
- Introduced LitMCPSpec for creating MCP tools, allowing for dynamic tool generation based on LitAPI endpoints.
- Updated LitServer to initialize MCP server with tools directly from the mcp_tools dictionary.
- Removed the obsolete mcp.py file to streamline the codebase.
- Improved error handling and logging in LitMCPSpec for better user feedback during tool creation.
